rating_feedback_service/requirements.txt
Django==4.2
python-decouple
djangorestframework
psycopg2-binary
